dhis2: implement `util` namespace (#1102)
* feat: create a util namespace


* feat: add changeset


* fix: remove skip from test and fix changeset


---------
