+++ packages/dhis2/src/index.js
@@ -4,3 +4,4 @@ export { metadata };
 import * as Adaptor from './Adaptor';
 export default Adaptor;
 export * from './Adaptor';
+export * as util from './util';
