+++ packages/dhis2/src/index.ts
@@ -8,3 +8,5 @@ import * as Adaptor from './Adaptor';
 export default Adaptor;
 
 export * from './Adaptor';
+
+export * as util from './util';
